+++ CHANGELOG.md
@@ -2,8 +2,13 @@
 
 ## [Unreleased]
 ### Added
+- Two new commit commands
+  - extend - add additional changes to previous commit
+  - amend - fix commit message
+- UI to display available commit commands
 - `--help` or `-h` flag for help information
 ### Changed
+- Use <kbd>c c</kbd> to create a commit
 - `-v` flag changed to `-V`
 - User is notified of unrecognised command line arguments and gex exits instead of quietly ignoring
 - Status says "Unstaged changes" and "Staged changes" instead of "files"